% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{as.list.llist}
\alias{as.list.llist}
\title{Translate a linked list into a list.}
\usage{
\method{as.list}{llist}(x, ...)
}
\arguments{
\item{x}{The linked list}
}
\value{
A \code{list} object containing the same elements as the linked list.
}
\description{
Translate a linked list into a list.
}
