% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{as.vector.llist}
\alias{as.vector.llist}
\title{Translate a linked list into a vector}
\usage{
\method{as.vector}{llist}(x, ...)
}
\arguments{
\item{x}{The linked list}
}
\value{
A vector object containing the same elements as the linked list.
}
\description{
Translate a linked list into a vector
}
