% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{CONS}
\alias{CONS}
\title{Constructs a new linked list from an element and another list}
\usage{
CONS(car, cdr)
}
\arguments{
\item{car}{The head of the new list}

\item{cdr}{The remainder of the list}
}
\value{
A new list
}
\description{
Constructs a new linked list from an element and another list
}
