% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{dequeue}
\alias{dequeue}
\title{Remove an element from the front of a queue}
\usage{
dequeue(queue)
}
\arguments{
\item{queue}{The queue}

\item{elm}{An element to add to the queue}
}
\value{
The modified queue
}
\description{
Remove an element from the front of a queue
}
