% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{empty_red_black_map}
\alias{empty_red_black_map}
\title{Create an empty red-black tree representation for a map}
\usage{
empty_red_black_map()
}
\description{
Create an empty red-black tree representation for a map
}
