% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{enqueue}
\alias{enqueue}
\title{Add an element to the back of a queue}
\usage{
enqueue(queue, elm)
}
\arguments{
\item{queue}{The queue}

\item{elm}{An element to add to the queue}
}
\value{
The modified queue
}
\description{
Add an element to the back of a queue
}
