% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{front}
\alias{front}
\title{Get the element at the front of a queue}
\usage{
front(queue)
}
\arguments{
\item{queue}{The queue}
}
\value{
a list containing two values, the element at the front of the queue
as well as the modified queue.
}
\description{
Get the element at the front of a queue
}
