% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{is_stack_empty}
\alias{is_stack_empty}
\title{Tests if a stack is empty}
\usage{
is_stack_empty(stack)
}
\arguments{
\item{stack}{A stack}
}
\value{
\code{TRUE} if \code{stack} is empty, \code{FALSE} otherwise
}
\description{
Tests if a stack is empty
}
