% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{ll_is_nil}
\alias{ll_is_nil}
\title{Tests if a list is empty.}
\usage{
ll_is_nil(llist)
}
\arguments{
\item{llist}{The linked list}
}
\value{
\code{TRUE} if \code{llist} is empty (i.e. equal to \code{NIL}) and \code{FALSE}
otherwise.
}
\description{
Tests if a list is empty.
}
