% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llconcat}
\alias{llconcat}
\title{Concatenates two linked lists.}
\usage{
llconcat(l1, l2, cnt = identity)
}
\arguments{
\item{l1}{First list.}

\item{l2}{Second list.}

\item{cnt}{Continuation to make the function tail-recursive}
}
\value{
The concatenation of \code{l1} and \code{l2}.
}
\description{
Concatenates two linked lists.
}
