% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llcontains}
\alias{llcontains}
\title{Tests if an element is contained in a list}
\usage{
llcontains(llist, elm)
}
\arguments{
\item{llist}{A linked list}

\item{elm}{An element}
}
\value{
\code{TRUE`` if}elm\code{is in}llist\code{and}FALSE` otherwise
}
\description{
Tests if an element is contained in a list
}
