% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{lldrop}
\alias{lldrop}
\title{Remove the first \code{k} elements from a list}
\usage{
lldrop(llist, k, acc = NIL)
}
\arguments{
\item{llist}{The linked list}

\item{k}{The number of elements to remove}
}
\value{
The list without the first k elements
}
\description{
Remove the first \code{k} elements from a list
}
