% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llength}
\alias{llength}
\title{Compute the length of a linked list.}
\usage{
llength(llist, acc = 0)
}
\arguments{
\item{llist}{The linked list}

\item{acc}{An accumulator (for making the function tail-recursive)}
}
\value{
The length of \code{llist}
}
\description{
This function computes the length of a linked list by scanning
through it. This is thus a linear time function.
}
