% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llfilter}
\alias{llfilter}
\title{Remove elements that do not satisfy a predicate}
\usage{
llfilter(llist, p, acc = NIL)
}
\arguments{
\item{llist}{The list}

\item{p}{A predicate function}

\item{acc}{Accumulator to make the function tail-recursive}
}
\description{
Remove elements that do not satisfy a predicate
}
