% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llist_from_list}
\alias{llist_from_list}
\title{Translate a list object into a linked list.}
\usage{
llist_from_list(x)
}
\arguments{
\item{x}{A \code{list} object}
}
\value{
A linked list containing the elements in \code{x}
}
\description{
Translate a list object into a linked list.
}
