% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llmap}
\alias{llmap}
\title{Map a function over a linked list.}
\usage{
llmap(llist, f, acc = NIL)
}
\arguments{
\item{llist}{The linked list}

\item{f}{The function to apply to all elements in \code{llist}}

\item{acc}{Accumulator to make the function tail-recursive}
}
\description{
Map a function over a linked list.
}
