% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llrev}
\alias{llrev}
\title{Reverse a linked list.}
\usage{
llrev(llist, acc = NIL)
}
\arguments{
\item{llist}{A linked list}

\item{acc}{An accumulator (for making the function tail-recursive)}
}
\value{
A list containing the elements in \code{llist} in reverse order.
}
\description{
This function scans through \code{llist}, building the reverse list in \code{acc}.
}
