% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{lltake}
\alias{lltake}
\title{Extract the first \code{k} elements from a list}
\usage{
lltake(llist, k, acc = NIL)
}
\arguments{
\item{llist}{The linked list}

\item{k}{The number of elements to take}

\item{acc}{Accumulator to make the function tail-recursive}
}
\value{
The first \code{k} elements of \code{llist}
}
\description{
Extract the first \code{k} elements from a list
}
