% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{move_lists}
\alias{move_lists}
\title{Move elements from the back list to the front list}
\usage{
move_lists(queue)
}
\arguments{
\item{queue}{A queue}
}
\value{
modified queue
}
\description{
Move elements from the back list to the front list
}
