% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{pop}
\alias{pop}
\title{Remove the top element from a stack}
\usage{
pop(stack)
}
\arguments{
\item{stack}{A non-empty stack}
}
\value{
A copy of \code{stack} where the first element has been removed
}
\description{
Remove the top element from a stack
}
