% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{print.llist}
\alias{print.llist}
\title{Print linked list objects}
\usage{
\method{print}{llist}(x, ...)
}
\arguments{
\item{x}{The linked list}

\item{...}{Additional parameters (not used in this version)}
}
\description{
Print linked list objects
}
