% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{print.rbt_colour}
\alias{print.rbt_colour}
\title{Print red-black colour objects}
\usage{
\method{print}{rbt_colour}(x, ...)
}
\arguments{
\item{x}{The colour}

\item{...}{Additional parameters (not used in this version)}
}
\description{
Print red-black colour objects
}
