% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{print.rbt_map}
\alias{print.rbt_map}
\title{Print red-black map objects}
\usage{
\method{print}{rbt_map}(x, ...)
}
\arguments{
\item{x}{The set}

\item{...}{Additional parameters (not used in this version)}
}
\description{
Print red-black map objects
}
