% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{print.rbt_set}
\alias{print.rbt_set}
\title{Print red-black sets objects}
\usage{
\method{print}{rbt_set}(x, ...)
}
\arguments{
\item{x}{The set}

\item{...}{Additional parameters (not used in this version)}
}
\description{
Print red-black sets objects
}
