% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{push}
\alias{push}
\title{Push an element onto a stack.}
\usage{
push(stack, elm)
}
\arguments{
\item{stack}{A stack object}

\item{elm}{An element to push to the top of the stack}
}
\value{
Updated stack
}
\description{
Push an element onto a stack.
}
