% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\docType{data}
\name{RBT_DOUBLE_BLACK}
\alias{RBT_DOUBLE_BLACK}
\title{Colour used in red-black trees.}
\format{An object of class \code{rbt_colours} of length 1.}
\usage{
RBT_DOUBLE_BLACK
}
\description{
Colour used in red-black trees.
}
\keyword{datasets}
