% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{RBT_MAP}
\alias{RBT_MAP}
\title{Constructs a new node in a red-black tree map.}
\usage{
RBT_MAP(col, key, val, left, right)
}
\arguments{
\item{col}{The colour for the node.}

\item{key}{The key for the map stored in the node}

\item{val}{The value kept in the node}

\item{left}{The left subtree}

\item{right}{The right subtree}
}
\value{
\preformatted{ A new tree node
}
}
\description{
Constructs a new node in a red-black tree map.
}
