% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\docType{data}
\name{RBT_MAP_EMPTY}
\alias{RBT_MAP_EMPTY}
\title{An empty map}
\format{An object of class \code{rbt_map} of length 1.}
\usage{
RBT_MAP_EMPTY
}
\description{
An empty map
}
\keyword{datasets}
