% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{rbt_map_get}
\alias{rbt_map_get}
\title{Get the value associated with a key.}
\usage{
rbt_map_get(tree, k)
}
\arguments{
\item{tree}{The red-black tree}

\item{k}{The key to search for}
}
\description{
Get the value associated with a key.
}
