% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{rbt_map_member}
\alias{rbt_map_member}
\title{Determines if a red-black map contains the key \code{k}}
\usage{
rbt_map_member(tree, k)
}
\arguments{
\item{tree}{The red-black tree}

\item{k}{The key to search for}
}
\description{
Determines if a red-black map contains the key \code{k}
}
