% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{RBT_SET}
\alias{RBT_SET}
\title{Constructs a new node in a red-black tree set.}
\usage{
RBT_SET(col, val, left, right)
}
\arguments{
\item{col}{The colour for the node.}

\item{val}{The value kept in the node}

\item{left}{The left subtree}

\item{right}{The right subtree}
}
\value{
\preformatted{ A new tree node
}
}
\description{
Constructs a new node in a red-black tree set.
}
