% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\docType{data}
\name{RBT_SET_EMPTY}
\alias{RBT_SET_EMPTY}
\title{An empty set}
\format{An object of class \code{rbt_set} of length 1.}
\usage{
RBT_SET_EMPTY
}
\description{
An empty set
}
\keyword{datasets}
