% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{rbt_set_insert}
\alias{rbt_set_insert}
\title{Insert an element into a red-black tree set.}
\usage{
rbt_set_insert(tree, elm)
}
\arguments{
\item{tree}{The set}

\item{elm}{The element to insert}
}
\description{
Insert an element into a red-black tree set.
}
