% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{rbt_set_member}
\alias{rbt_set_member}
\title{Determines if a red-black tree contains the value \code{v}}
\usage{
rbt_set_member(tree, v)
}
\arguments{
\item{tree}{The red-black tree}

\item{v}{The value to search for}
}
\description{
Determines if a red-black tree contains the value \code{v}
}
