% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{top}
\alias{top}
\title{Get the element at the top of the stack.}
\usage{
top(stack)
}
\arguments{
\item{stack}{A non-empty stack}
}
\value{
The element at the top of \code{stack}
}
\description{
Get the element at the top of the stack.
}
