% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{toString.llist}
\alias{toString.llist}
\title{Make a string-representation of a linked list.}
\usage{
\method{toString}{llist}(x, ...)
}
\arguments{
\item{x}{The linked list}

\item{...}{Additional parameters (not used in this version)}
}
\value{
A string representation of the list
}
\description{
Make a string-representation of a linked list.
}
