% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{toString.rbt_colour}
\alias{toString.rbt_colour}
\title{Make a string-representation of a red-black colour}
\usage{
\method{toString}{rbt_colour}(x, ...)
}
\arguments{
\item{x}{The set}

\item{...}{Additional parameters (not used in this version)}
}
\value{
A string representation of the colour
}
\description{
Make a string-representation of a red-black colour
}
