% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{toString.rbt_map}
\alias{toString.rbt_map}
\title{Make a string-representation of a red-black map.}
\usage{
\method{toString}{rbt_map}(x, ...)
}
\arguments{
\item{x}{The map}

\item{...}{Additional parameters (not used in this version)}
}
\value{
A string representation of the map
}
\description{
Make a string-representation of a red-black map.
}
