% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{toString.rbt_set}
\alias{toString.rbt_set}
\title{Make a string-representation of a red-black set.}
\usage{
\method{toString}{rbt_set}(x, ...)
}
\arguments{
\item{x}{The set}

\item{...}{Additional parameters (not used in this version)}
}
\value{
A string representation of the set
}
\description{
Make a string-representation of a red-black set.
}
