% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{empty_red_black_map}
\alias{empty_red_black_map}
\title{Create an empty red-black tree representation for a map}
\usage{
empty_red_black_map()
}
\description{
While we do have \code{RBT_SET_EMPTY} to represent an empty tree, the rotations are
simpler to implement if leaves are actually inner nodes with colours and just not
containing any value.
}
