% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red-black-trees.R
\name{empty_red_black_set}
\alias{empty_red_black_set}
\title{Create an empty red-black tree representation for a set.}
\usage{
empty_red_black_set()
}
\description{
Create an empty red-black tree representation for a set.
}
