% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{llcontains}
\alias{llcontains}
\title{Checks if a linked list contains an element}
\usage{
llcontains(llist, elm)
}
\arguments{
\item{llist}{The linked list}

\item{elm}{The element}
}
\value{
\code{TRUE} if \code{elm} is in \code{llist} and \code{FALSE} otherwise
}
\description{
Checks if a linked list contains an element
}
