% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-lists.R
\name{lltake}
\alias{lltake}
\title{Extract the first \code{k} elements from a linked list.}
\usage{
lltake(llist, k, acc = NIL)
}
\arguments{
\item{llist}{A linked list}

\item{k}{The number of elements to extract}
}
\value{
A new linked list containing the first \code{k} elements of \code{llist}.
}
\description{
Extract the first \code{k} elements from a linked list.
}
